backend/pyproject.toml
[project]
name = "task-management-backend"
version = "0.1.0"
description = "Backend for task management application with authentication"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.20",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "bcrypt>=5.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.1.0",
    "mypy>=1.13.0",
]
